refactor: split vga stuff into multiple files
