+++ src/vga/color.rs
@@ -1,0 +1,31 @@
+#[repr(u8)]
+#[allow(unused)]
+#[derive(Debug, Clone, Copy)]
+pub enum Color {
+    Black = 0,
+    Blue = 1,
+    Green = 2,
+    Cyan = 3,
+    Red = 4,
+    Magenta = 5,
+    Brown = 6,
+    LightGrey = 7,
+    DarkGrey = 8,
+    LightBlue = 9,
+    LightGreen = 10,
+    LightCyan = 11,
+    LightRed = 12,
+    LightMagenta = 13,
+    LightBrown = 14,
+    White = 15,
+}
+
+impl Color {
+    pub fn to_foreground(self) -> u8 {
+        self as u8
+    }
+
+    pub fn to_background(self) -> u8 {
+        (self as u8) << 4
+    }
+}
