+++ src/vga/mod.rs
@@ -1,39 +1,12 @@
-use core::{arch::asm, ptr::write_volatile};
+use crate::vga::{buffer::Buffer, color::Color, cursor::Cursor};
+use core::ptr::write_volatile;
+
+mod buffer;
+mod color;
+mod cursor;
 
 #[cfg(test)]
 use spin::Mutex;
-
-#[repr(u8)]
-#[allow(unused)]
-#[derive(Debug, Clone, Copy)]
-pub enum Color {
-    Black = 0,
-    Blue = 1,
-    Green = 2,
-    Cyan = 3,
-    Red = 4,
-    Magenta = 5,
-    Brown = 6,
-    LightGrey = 7,
-    DarkGrey = 8,
-    LightBlue = 9,
-    LightGreen = 10,
-    LightCyan = 11,
-    LightRed = 12,
-    LightMagenta = 13,
-    LightBrown = 14,
-    White = 15,
-}
-
-impl Color {
-    pub fn to_foreground(self) -> u8 {
-        self as u8
-    }
-
-    pub fn to_background(self) -> u8 {
-        (self as u8) << 4
-    }
-}
 
 pub const VGA_WIDTH: u8 = 80;
 pub const VGA_HEIGHT: u8 = 25;
@@ -61,136 +34,10 @@ pub enum Direction {
     Right,
 }
 
+#[cfg(test)]
+static VGA_BUFFER_LOCK: Mutex<()> = Mutex::new(());
+
 pub struct OutOfBoundsError;
-
-/// Abstraction for managing the [Text-mode cursor](https://wiki.osdev.org/Text_Mode_Cursor).
-#[derive(Clone, Copy)]
-pub struct Cursor {}
-
-#[allow(unused)]
-impl Cursor {
-    const LOCATION_REG_LOW: u8 = 0x0F;
-    const LOCATION_REG_HIGH: u8 = 0x0E;
-    const REG_START: u8 = 0x0A;
-    const REG_END: u8 = 0x0B;
-
-    /// Updates the text-mode cursor position in the VGA buffer by setting the CRTC's
-    /// [location registers](http://www.osdever.net/FreeVGA/vga/crtcreg.htm#0F) (`0x0F` and `0x0D`)
-    /// to `x, y`.
-    ///
-    /// ## SAFETY
-    /// 1.  This function uses `Cursor::update`, which writes directly to the VGA buffer. In user-mode, this **will** result
-    ///     in invalid memory access.
-    ///
-    /// 2.  `update_pos` may cause undefined behavior if called with `x` or `y` values outside of the range `0x00..=0x0F`.
-    pub unsafe fn update_pos(&self, x: u16, y: u16) {
-        let out_of_bounds: bool = !(0..VGA_HEIGHT).contains(&(y as u8)) || !(0..VGA_WIDTH).contains(&(x as u8));
-        if out_of_bounds {
-            return;
-        }
-
-        let pos = y * VGA_WIDTH as u16 + x;
-
-        self.update(Cursor::LOCATION_REG_LOW, (pos & 0xFF) as u8);
-        self.update(Cursor::LOCATION_REG_HIGH, ((pos >> 8) & 0xFF) as u8);
-    }
-
-    /// Resizes the cursor by updating the [cursor end & start register](http://www.osdever.net/FreeVGA/vga/crtcreg.htm#0A)
-    /// (`0x0A` and `0x0B`) to `start, end`. The values of `start` and `end` are expected to be in the range `0x00..=0x0F`.
-    ///
-    /// ## SAFETY
-    /// 1.  This function uses `Cursor::update`, which writes directly to the VGA buffer. In user-mode, this **will** result
-    ///     in invalid memory access.
-    ///
-    /// 2.  `resize` may cause undefined behavior if called with `start` or `end` values outside of the range `0x00..=0x0F`.
-    pub unsafe fn resize(&self, start: u8, end: u8) {
-        self.update(Cursor::REG_START, start);
-        self.update(Cursor::REG_END, end);
-    }
-
-    /// Abstraction for the ugliness behind updating the cursor.
-    ///
-    /// `0x3D4` is the I/O port address for the VGA's CRTC ([Cathode-ray tube](https://en.wikipedia.org/wiki/Cathode-ray_tube))'s
-    /// index register. The value being loaded into it defines which CRTC functionality we want to access.
-    /// The different indices that can be loaded into it are documented [here](http://www.osdever.net/FreeVGA/vga/).
-    ///
-    /// After the index has been loaded into the `0x3D4`, `dx`, (where the index register is stored) can be incremented by
-    /// one. This will move it to `0x3D5`, the CRTC's data register, signifying the CRTC's readiness to receive the input values.
-    ///
-    /// ## SAFETY:
-    /// This writes to the VGA buffer directly, running this in a non-bare-metal environment
-    /// will result in invalid memory access.
-    unsafe fn update(&self, index: u8, value: u8) {
-        asm!(
-            "mov dx, 0x3D4",
-            "mov al, {index}",
-            "out dx, al",
-            "inc dx",
-            "mov al, {value}",
-            "out dx, al",
-            index = in(reg_byte) (index),
-            value = in(reg_byte) (value),
-            out("dx") _,
-            out("al") _,
-        )
-    }
-}
-
-#[derive(Clone, Copy)]
-/// Buffer implementation for storing content beyond the VGA buffer size of 4000 bytes (80 x 25
-/// u16 entries).
-///
-/// Intended to be used in the `Vga` implementation. Can be written to, and flushed into the VGA buffer
-/// using `Vga::flush`.
-pub struct Buffer {
-    buf: [u16; (VGA_WIDTH as usize) * (MAX_BUFFERED_LINES as usize)],
-}
-
-impl Buffer {
-    const NEWLINE: u8 = 0xFF;
-
-    /// Returns a new `Buffer` object with a buffer size of `VGA_WIDTH * MAX_BUFFERED_LINES`. `VGA_WIDTH`
-    /// is fixed to 80 (hardware limitation), and `MAX_BUFFERED_LINES` can be increased freely as long as enough memory
-    /// is available.
-    pub fn new() -> Self {
-        Self {
-            buf: [0u16; (VGA_WIDTH as usize) * (MAX_BUFFERED_LINES as usize)],
-        }
-    }
-
-    /// Writes `entry` to `self.buf[line_offset * VGA_WIDTH + rel_index]`.
-    /// ### Note
-    /// This does **not** write to the VGA buffer, only to the internal one. Writes to the VGA buffer are to be handled
-    /// by `Vga::flush`.
-    pub fn write(&mut self, line_offset: u8, rel_index: u16, entry: u16) {
-        let abs_index: usize = ((line_offset as usize * VGA_WIDTH as usize) + rel_index as usize) % self.buf.len();
-
-        self.buf[abs_index] = entry;
-    }
-
-    /// Returns a `self.buf` slice of `VGA_BUFFER_SIZE` starting at `line_offset`.
-    pub fn slice(&self, line_offset: u8) -> &[u16] {
-        let start = (line_offset as usize) * (VGA_WIDTH as usize);
-        let end = start + VGA_BUFFER_SIZE as usize;
-
-        &self.buf[start..end]
-    }
-
-    pub fn at(&self, pos: u16) -> Option<&u16> {
-        self.buf.get(pos as usize)
-    }
-
-    /// Returns the length of the written content starting from `from_x, from_y`, until either
-    /// the next newline, or if no newline is found, until the next null VGA entry, i.e the next
-    /// entry for which `(x & 0xFF) == 0` is true.
-    pub fn block_length(&self, from_x: u8, from_y: u8) -> u16 {
-        let slice = &self.buf[from_y as usize * VGA_WIDTH as usize + from_x as usize..];
-        if let Some(ind) = slice.iter().position(|x| *x == Buffer::NEWLINE as u16) {
-            return ind as u16;
-        }
-        slice.iter().position(|x| (*x & 0xFF) == 0).unwrap() as u16
-    }
-}
 
 #[derive(Clone, Copy)]
 /// Abstraction for VGA buffer interactions.
@@ -257,6 +104,17 @@ impl Vga {
     }
 
     /// Deletes the character from the VGA buffer at `self.x, self.y` and decrements the cursor.
+    ///
+    /// ### Example Usage:
+    /// ```
+    /// let mut v = Vga::new();
+    ///
+    /// loop {
+    ///     if let Some(char) = ps2::read_if_ready(None) == ps2::BACKSPACE {
+    ///         v.delete_char();
+    ///     }
+    /// }
+    /// ```
     pub fn delete_char(&mut self) {
         self.dec_cursor();
 
@@ -266,6 +124,13 @@ impl Vga {
 
     #[allow(unused)]
     /// Writes `s` to the VGA buffer starting at `self.x, self.y` and increments the cursor by `s.len()`.
+    ///
+    /// ### Example Usage
+    /// ```
+    /// let mut v = Vga::new();
+    ///
+    /// v.write_u8_arr(b"Hello, World!");
+    /// ```
     pub fn write_u8_arr(&mut self, s: &[u8]) {
         for c in s.iter() {
             if *c == 0 {
@@ -315,6 +180,13 @@ impl Vga {
 
     /// Sets the 4 most significant bits of `self.color` to `foreground`, setting the
     /// font color of the VGA buffer.
+    ///
+    /// ### Example Usage:
+    /// ```
+    /// let mut v = Vga::new();
+    ///
+    /// v.set_foreground_color(Vga::Color::White);
+    /// ```
     pub fn set_foreground_color(&mut self, foreground: Color) {
         self.color &= 0xF0;
         self.color |= foreground.to_foreground();
@@ -356,6 +228,7 @@ impl Vga {
         }
     }
 
+    /// Flushes `self.buffer.slice(self.line_offset)` into the VGA buffer, writing it to the console.
     fn flush(&self) {
         let current_displayed_content = self.buffer.slice(self.line_offset);
 
@@ -368,13 +241,18 @@ impl Vga {
 
     /// Writes `character` at `self.x == x` and `self.y == y` into the VGA buffer.
     fn write_char_at(&mut self, x: u8, y: u8, character: u8) -> Result<(), OutOfBoundsError> {
+        if y == VGA_HEIGHT - 1 && x == VGA_WIDTH - 1 {
+            return Ok(());
+        }
+
         if y >= VGA_HEIGHT || x >= VGA_WIDTH {
             return Err(OutOfBoundsError);
         }
-        let entry: u16 = (character as u16) | (self.color as u16) << 8;
-        let index: isize = ((y as usize + self.line_offset as usize) % MAX_BUFFERED_LINES as usize) as isize * VGA_WIDTH as isize + x as isize;
 
-        self.buffer.write(0, index as u16, entry);
+        let entry: u16 = (character as u16) | (self.color as u16) << 8;
+        let index: u16 = (y as u16 + self.line_offset as u16) * VGA_WIDTH as u16 + x as u16;
+
+        self.buffer.write(0, index, entry);
 
         Ok(())
     }
@@ -402,11 +280,15 @@ impl Vga {
 
     /// Increments the cursor, taking line wrapping into account.
     fn inc_cursor(&mut self) {
-        self.x += 1;
+        if self.x == VGA_WIDTH - 1 && self.y == VGA_HEIGHT - 1 {
+            return;
+        }
+
+        self.x = self.x.saturating_add(1);
 
         if self.x >= VGA_WIDTH {
             self.x = 0;
-            self.y += 1;
+            self.y = self.y.saturating_add(1);
         }
 
         if self.y >= VGA_HEIGHT {
@@ -420,6 +302,10 @@ impl Vga {
     }
 
     fn scroll_down(&mut self) {
+        if self.line_offset == MAX_BUFFERED_LINES - VGA_HEIGHT {
+            return;
+        }
+
         self.line_offset = (self.line_offset + 1).min(MAX_BUFFERED_LINES - VGA_HEIGHT);
         self.y = VGA_HEIGHT - 1;
 
@@ -433,12 +319,9 @@ impl Vga {
     /// Gets the position of the last written character for `y`, to ensure the cursor returns
     /// to the correct position when backspacing at `x == 0`.
     fn get_x_for_y(&self, y: usize) -> usize {
-        (self.buffer.block_length(0, (y + self.line_offset as usize) as u8).saturating_sub(1)) as usize
+        self.buffer.block_length(0, (y + self.line_offset as usize) as u8).saturating_sub(1) as usize
     }
 }
-
-#[cfg(test)]
-static VGA_BUFFER_LOCK: Mutex<()> = Mutex::new(());
 
 #[cfg(test)]
 mod test {
@@ -541,7 +424,7 @@ mod test {
         v.new_line();
 
         assert_eq!(
-            (v.buffer.buf[0] & 0xFF) as u8,
+            (v.buffer.at(0).unwrap() & 0xFF) as u8,
             b'H',
             "First character of the previous line should not be deleted when pressing enter"
         );
@@ -560,9 +443,9 @@ mod test {
         v.new_line();
 
         assert_eq!(
-            (v.buffer.buf[0] & 0xFF) as u8,
+            (v.buffer.at(0).unwrap() & 0xFF) as u8,
             b'H',
-            "First character of the previous line should not be deleted when pressing enter"
+            "First charactqer of the previous line should not be deleted when pressing enter"
         );
 
         v.clear_screen();
